refactor out sprites and add an alphabet
