+++ gitfiti.py
@@ -29,6 +29,9 @@ except NameError:
     # Python 3 (Python 2's `raw_input` was renamed to `input`)
     raw_input = input
 
+from sprites import IMAGES
+from sprites import LETTERS
+
 
 GITHUB_BASE_URL = 'https://github.com/'
 FALLBACK_IMAGE = 'kitty'
@@ -42,160 +45,6 @@ TITLE = '''
  \__, /_/\__/_/ /_/\__/_/
 /____/
 '''
-
-
-KITTY = [
-  [0,0,0,4,0,0,0,0,4,0,0,0],
-  [0,0,4,2,4,4,4,4,2,4,0,0],
-  [0,0,4,2,2,2,2,2,2,4,0,0],
-  [2,2,4,2,4,2,2,4,2,4,2,2],
-  [0,0,4,2,2,3,3,2,2,4,0,0],
-  [2,2,4,2,2,2,2,2,2,4,2,2],
-  [0,0,0,3,4,4,4,4,3,0,0,0],
-]
-
-ONEUP = [
-  [0,4,4,4,4,4,4,4,0],
-  [4,3,2,2,1,2,2,3,4],
-  [4,2,2,1,1,1,2,2,4],
-  [4,3,4,4,4,4,4,3,4],
-  [4,4,1,4,1,4,1,4,4],
-  [0,4,1,1,1,1,1,4,0],
-  [0,0,4,4,4,4,4,0,0],
-]
-
-ONEUP2 = [
-  [0,0,4,4,4,4,4,4,4,0,0],
-  [0,4,2,2,1,1,1,2,2,4,0],
-  [4,3,2,2,1,1,1,2,2,3,4],
-  [4,3,3,4,4,4,4,4,3,3,4],
-  [0,4,4,1,4,1,4,1,4,4,0],
-  [0,0,4,1,1,1,1,1,4,0,0],
-  [0,0,0,4,4,4,4,4,0,0,0],
-]
-
-HACKERSCHOOL = [
-  [4,4,4,4,4,4],
-  [4,3,3,3,3,4],
-  [4,1,3,3,1,4],
-  [4,3,3,3,3,4],
-  [4,4,4,4,4,4],
-  [0,0,4,4,0,0],
-  [4,4,4,4,4,4],
-]
-
-OCTOCAT = [
-  [0,0,0,4,0,0,0,4,0],
-  [0,0,4,4,4,4,4,4,4],
-  [0,0,4,1,3,3,3,1,4],
-  [4,0,3,4,3,3,3,4,3],
-  [0,4,0,0,4,4,4,0,0],
-  [0,0,4,4,4,4,4,4,4],
-  [0,0,4,0,4,0,4,0,4],
-]
-
-OCTOCAT2 = [
-  [0,0,4,0,0,4,0],
-  [0,4,4,4,4,4,4],
-  [0,4,1,3,3,1,4],
-  [0,4,4,4,4,4,4],
-  [4,0,0,4,4,0,0],
-  [0,4,4,4,4,4,0],
-  [0,0,0,4,4,4,0],
-]
-
-HELLO = [
-  [0,1,0,0,0,0,0,0,0,1,0,1,0,0,0,0,0,4],
-  [0,2,0,0,0,0,0,0,0,2,0,2,0,0,0,0,0,4],
-  [0,3,3,3,0,2,3,3,0,3,0,3,0,1,3,1,0,3],
-  [0,4,0,4,0,4,0,4,0,4,0,4,0,4,0,4,0,3],
-  [0,3,0,3,0,3,3,3,0,3,0,3,0,3,0,3,0,2],
-  [0,2,0,2,0,2,0,0,0,2,0,2,0,2,0,2,0,0],
-  [0,1,0,1,0,1,1,1,0,1,0,1,0,1,1,1,0,4],
-]
-
-HIREME = [
-  [1,0,0,0,4,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],
-  [2,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],
-  [3,3,3,0,2,0,3,3,3,0,2,3,3,0,0,3,3,0,3,0,0,2,3,3],
-  [4,0,4,0,4,0,4,0,0,0,4,0,4,0,0,4,0,4,0,4,0,4,0,4],
-  [3,0,3,0,3,0,3,0,0,0,3,3,3,0,0,3,0,3,0,3,0,3,3,3],
-  [2,0,2,0,2,0,2,0,0,0,2,0,0,0,0,2,0,2,0,2,0,2,0,0],
-  [1,0,1,0,1,0,1,0,0,0,1,1,1,0,0,1,0,1,0,1,0,1,1,1],
-]
-
-BEER = [
-  [0,0,0,0,0,0,0,3,3,3,0,0,3,3,3,0,3,3,3,0,3,3,3,0,0],
-  [0,0,1,1,1,1,0,3,0,0,3,0,3,0,0,0,3,0,0,0,3,0,0,3,0],
-  [0,2,2,2,2,2,0,3,0,0,3,0,3,0,0,0,3,0,0,0,3,0,0,3,0],
-  [2,0,2,2,2,2,0,3,3,3,0,0,3,3,3,0,3,3,3,0,3,3,3,0,0],
-  [2,0,2,2,2,2,0,3,0,0,3,0,3,0,0,0,3,0,0,0,3,0,3,0,0],
-  [0,2,2,2,2,2,0,3,0,0,3,0,3,0,0,0,3,0,0,0,3,0,0,3,0],
-  [0,0,2,2,2,2,0,3,3,3,0,0,3,3,3,0,3,3,3,0,3,0,0,3,0],
-]
-
-GLIDERS = [
-  [0,0,0,4,0,4,0,0,0,0,4,0,0,0],
-  [0,4,0,4,0,0,4,4,0,0,0,4,0,0],
-  [0,0,4,4,0,4,4,0,0,4,4,4,0,0],
-  [0,0,0,0,0,0,0,0,0,0,0,0,0,0],
-  [0,4,0,4,0,0,0,4,0,0,0,0,0,0],
-  [0,0,4,4,0,4,0,4,0,0,0,0,0,0],
-  [0,0,4,0,0,0,4,4,0,0,0,0,0,0],
-]
-
-ASCII_TO_NUMBER = {
-  '_': 0,
-  '_': 1,
-  '~': 2,
-  '=': 3,
-  '*': 4,
-}
-
-
-def str_to_sprite(content):
-    # Break out lines and filter any excess
-    lines = content.split('\n')
-    def is_empty_line(line):
-        return len(line) != 0
-    lines = filter(is_empty_line, lines)
-
-    # Break up lines into each character
-    split_lines = [list(line) for line in lines]
-
-    # Replace each character with its numeric equivalent
-    for line in split_lines:
-        for index, char in enumerate(line):
-            line[index] = ASCII_TO_NUMBER.get(char, 0)
-
-    # Return the formatted str
-    return split_lines
-
-
-ONEUP_STR = str_to_sprite('''
- *******
-*=~~-~~=*
-*~~---~~*
-*=*****=*
-**-*-*-**
- *-----*
-  *****
-''')
-
-
-IMAGES = {
-  'kitty': KITTY,
-  'oneup': ONEUP,
-  'oneup2': ONEUP2,
-  'hackerschool': HACKERSCHOOL,
-  'octocat': OCTOCAT,
-  'octocat2': OCTOCAT2,
-  'hello': HELLO,
-  'hireme': HIREME,
-  'oneup_str': ONEUP_STR,
-  'beer': BEER,
-  'gliders': GLIDERS,
-}
 
 
 def load_images(img_names):
@@ -401,6 +250,8 @@ def main():
     if not image:
         image = IMAGES[image_name_fallback]
     else:
+        if len(image) == 1:
+            image = LETTERS[image]
         try:
             image = images[image]
         except:
